Кнопку вынес в отдельный компонент т.к. встречается в трех местах и есть логика на нее
